Refactor: Endpoint update book info and fix update selects
Added new methods to book_data_methods for recieving a specific book, removing book data and moving a book to a folder (last 2 wip), made edit_book_data acknowledge new_folder and new title values.
Rigged up the rename_book endpoint as a prototype for having the endpoints update book_data.
Updated update management selects to work with the new book data files(thumbs still need work due to lack of integrated endpoints)
small improvements(swaped some lets to const and removed book map) in populate_management_selects.js

+++ Assets/Scripts/API_segments/Book_methods.py
@@ -1,4 +1,5 @@
 import os
+from .Book_data_methods import edit_book_data, get_specific_book_data, remove_book_data
 
 
 def list_books(mainDir):
@@ -31,12 +32,16 @@ def Remove_book(book_name, ext, folder, mainDir):
     target = "{0}/{1}/{2}.{3}".format(mainDir, folder, book_name, ext)
     if os.path.exists(target):
         try:
-            os.remove(target)
+            result = remove_book_data(folder, book_name)
+            if result == "200":
+                os.remove(target)
+                return result
+            else:
+                raise Exception('Failed to remove book')
         except Exception as e:
             return "500: " + str(e)
     else:
         return "404"
-    return "200"
 
 
 def Rename_book(book_name, ext, folder, new_name, mainDir):
@@ -49,10 +54,20 @@ def Rename_book(book_name, ext, folder, new_name, mainDir):
         return "409"
     if os.path.exists(target):
         try:
-            os.rename(
-                target, renamed)
+            # try and update the book data, rename if successfull, revert if not
+            data = get_specific_book_data(folder, book_name)
+            result = edit_book_data(folder, book_name, data["Authors"], data["Date"], data["Publisher"],
+                                    data["isbn"], data["isbn13"], data["Thumbnail"], data["Extension"], "True", "", new_name)
+            if result == "200":
+                os.rename(
+                    target, renamed)
+            else:
+                edit_book_data(folder, book_name, data["Authors"], data["Date"], data["Publisher"],
+                               data["isbn"], data["isbn13"], data["Thumbnail"], data["Extension"], "True")
+                return "400"
+
+            return result
         except Exception as e:
             return "500: " + str(e)
     else:
         return "404"
-    return "200"
